# internal/api/handlers_xiaomi.go
package api

import (
	"context"
	"encoding/base64"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gorilla/websocket"
	"github.com/lalmax-pro/lalmax-nvr/internal/camera"
	"github.com/lalmax-pro/lalmax-nvr/internal/config"
	"github.com/lalmax-pro/lalmax-nvr/internal/xiaomi"
)

// --- Xiaomi cloud endpoints ---

func (h *Handler) handleXiaomiAuth(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
		return
	}

	var req struct {
		Username string `json:"username"`
		Password string `json:"password"`
		Region   string `json:"region,omitempty"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid request body")
		return
	}
	if req.Username == "" || req.Password == "" {
		writeError(w, http.StatusBadRequest, "username and password are required")
		return
	}

	region := req.Region
	if region == "" {
		region = "cn"
	}

	result, verification, err := h.cloudProxy.SignIn(r.Context(), req.Username, req.Password, region)
	if err != nil {
		writeError(w, http.StatusUnauthorized, fmt.Sprintf("authentication failed: %v", err))
		return
	}

	if verification != nil {
		writeJSON(w, http.StatusAccepted, verificationToResponse(verification))
		return
	}

	// Store token in config
	h.saveXiaomiToken(result)

	writeJSON(w, http.StatusOK, map[string]any{
		"status":  "ok",
		"user_id": result.UserID,
	})
}

func (h *Handler) handleXiaomiCaptcha(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
		return
	}

	var req struct {
		SessionID   string `json:"session_id"`
		CaptchaCode string `json:"captcha_code"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid request body")
		return
	}
	if req.SessionID == "" || req.CaptchaCode == "" {
		writeError(w, http.StatusBadRequest, "session_id and captcha_code are required")
		return
	}

	result, verification, err := h.cloudProxy.SubmitCaptcha(r.Context(), req.SessionID, req.CaptchaCode)
	if err != nil {
		writeError(w, http.StatusUnauthorized, fmt.Sprintf("captcha verification failed: %v", err))
		return
	}

	if verification != nil {
		writeJSON(w, http.StatusAccepted, verificationToResponse(verification))
		return
	}

	// Store token in config
	h.saveXiaomiToken(result)

	writeJSON(w, http.StatusOK, map[string]any{
		"status":  "ok",
		"user_id": result.UserID,
	})
}

func (h *Handler) handleXiaomiVerify(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
		return
	}

	var req struct {
		SessionID string `json:"session_id"`
		Ticket    string `json:"ticket"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "invalid request body")
		return
	}
	if req.SessionID == "" || req.Ticket == "" {
		writeError(w, http.StatusBadRequest, "session_id and ticket are required")
		return
	}

	result, verification, err := h.cloudProxy.SubmitVerify(r.Context(), req.SessionID, req.Ticket)
	if err != nil {
		writeError(w, http.StatusUnauthorized, fmt.Sprintf("verification failed: %v", err))
		return
	}

	if verification != nil {
		writeJSON(w, http.StatusAccepted, verificationToResponse(verification))
		return
	}

	// Store token in config
	h.saveXiaomiToken(result)

	writeJSON(w, http.StatusOK, map[string]any{
		"status":  "ok",
		"user_id": result.UserID,
	})
}

func (h *Handler) handleXiaomiDevices(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
		return
	}

	// Get stored token from config
	if h.config == nil || h.config.Xiaomi.Token == "" {
		writeJSON(w, http.StatusOK, map[string]any{
			"devices": []CloudDeviceInfo{},
			"message": "not authenticated",
		})
		return
	}

	devices, err := h.cloudProxy.ListDevices(r.Context())
	if err != nil {
		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to get devices: %v", err))
		return
	}

	writeJSON(w, http.StatusOK, map[string]any{
		"devices": devices,
	})
}

// handleXiaomiSync syncs Xiaomi cloud device info (name, model, MAC) to NVR camera config.
// It fetches all devices from Xiaomi cloud, matches them to existing NVR cameras by DID,
// and updates name, model, brand, and serial_number (MAC). Returns count of synced cameras.
func (h *Handler) handleXiaomiSync(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
		return
	}
	if h.config == nil || h.config.Xiaomi.Token == "" {
		writeError(w, http.StatusUnauthorized, "xiaomi cloud not authenticated")
		return
	}
	if h.camMgr == nil {
		writeError(w, http.StatusInternalServerError, "camera manager not available")
		return
	}

	devices, err := h.cloudProxy.ListDevices(r.Context())
	if err != nil {
		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to get devices: %v", err))
		return
	}

	// Build DID → CloudDeviceInfo lookup
	deviceByDID := make(map[string]*CloudDeviceInfo, len(devices))
	for i := range devices {
		deviceByDID[devices[i].DID] = &devices[i]
	}

	synced := 0
	for i := range h.config.Cameras {
		cam := &h.config.Cameras[i]
		if cam.Protocol != "xiaomi" {
			continue
		}

		did := cam.DID
		if did == "" {
			did = extractDIDFromURL(cam.URL)
		}
		if did == "" {
			continue
		}

		dev, ok := deviceByDID[did]
		if !ok {
			continue
		}

		updates := camera.CameraUpdate{
			Name:         &dev.Name,
			Model:        &dev.Model,
			Brand:        strPtr("Xiaomi"),
			SerialNumber: &dev.MAC,
		}

		if _, err := h.camMgr.UpdateCamera(context.Background(), cam.ID, updates); err != nil {
			logger.Warn("failed to sync xiaomi camera", "camera_id", cam.ID, "did", did, "error", err)
			continue
		}
		synced++
	}

	writeJSON(w, http.StatusOK, map[string]any{
		"synced": synced,
		"total":  len(deviceByDID),
	})
}

func (h *Handler) handleCheckVendor(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeJSON(w, http.StatusOK, map[string]any{
			"vendor":     "unknown",
			"compatible": true,
		})
		return
	}
	if h.config == nil || h.config.Xiaomi.Token == "" {
		writeJSON(w, http.StatusOK, map[string]any{
			"vendor":     "unknown",
			"compatible": true,
		})
		return
	}

	did := r.URL.Query().Get("did")
	if did == "" {
		writeError(w, http.StatusBadRequest, "did parameter required")
		return
	}

	vendor, err := h.cloudProxy.CheckVendor(r.Context(), did)
	if err != nil {
		// For errors, return unknown/compatible (don't block on uncertainty)
		writeJSON(w, http.StatusOK, map[string]any{
			"vendor":     "unknown",
			"compatible": true,
		})
		return
	}

	if vendor == "tutk" {
		writeJSON(w, http.StatusOK, map[string]any{
			"vendor":     "tutk",
			"compatible": false,
			"message":    "This device uses TUTK protocol which is not supported by lalmax-nvr. Only CS2 protocol cameras are supported.",
		})
		return
	}

	writeJSON(w, http.StatusOK, map[string]any{
		"vendor":     "cs2",
		"compatible": true,
	})
}

// saveXiaomiToken persists auth result to config file.
func (h *Handler) saveXiaomiToken(result *CloudAuthResult) {
	if h.config == nil || result == nil {
		return
	}
	h.config.Xiaomi.UserID = result.UserID
	h.config.Xiaomi.Token = result.PassToken
	h.config.Xiaomi.Region = result.Region
	if err := config.Save(h.configPath, h.config); err != nil {
		logger.Warn("failed to save xiaomi config", "error", err)
	}

	// Also push to the cloud proxy so it has the latest credentials
	if h.cloudProxy != nil {
		_ = h.cloudProxy.SetCloudConfig(context.Background(), result.UserID, result.PassToken, result.Region)
	}
}

// verificationToResponse converts a CloudVerificationRequired to an API response map.
func verificationToResponse(v *CloudVerificationRequired) map[string]any {
	resp := map[string]any{
		"status": "verification_required",
	}
	if len(v.Captcha) > 0 {
		resp["captcha"] = base64.StdEncoding.EncodeToString(v.Captcha)
	}
	if v.VerifyPhone != "" {
		resp["verify_phone"] = v.VerifyPhone
	}
	if v.VerifyEmail != "" {
		resp["verify_email"] = v.VerifyEmail
	}
	if v.CaptchaSessionID != "" {
		resp["session_id"] = v.CaptchaSessionID
	}
	return resp
}

// HandleXiaomiTalkWS handles WebSocket connections for Xiaomi camera two-way audio.
// Browser sends PCMA audio via WebSocket, server forwards to camera speaker via MISS protocol.
func (h *Handler) handleXiaomiTalkWS(w http.ResponseWriter, r *http.Request) {
	if h.cloudProxy == nil {
		writeError(w, http.StatusServiceUnavailable, "xiaomi cloud not available")
		return
	}

	cameraID := r.URL.Query().Get("camera_id")
	if cameraID == "" {
		writeError(w, http.StatusBadRequest, "camera_id is required")
		return
	}

	// Find camera config
	var cam *config.CameraConfig
	for i := range h.config.Cameras {
		if h.config.Cameras[i].ID == cameraID {
			cam = &h.config.Cameras[i]
			break
		}
	}
	if cam == nil {
		writeError(w, http.StatusNotFound, "camera not found")
		return
	}
	if cam.Protocol != "xiaomi" {
		writeError(w, http.StatusBadRequest, "camera is not a xiaomi camera")
		return
	}

	// Resolve MISS URL
	did := cam.DID
	if did == "" {
		did = extractDIDFromURL(cam.URL)
	}
	if did == "" {
		writeError(w, http.StatusBadRequest, "camera has no DID")
		return
	}

	cloudCfg := xiaomi.XiaomiCloudConfig{
		UserID: h.config.Xiaomi.UserID,
		Token:  h.config.Xiaomi.Token,
		Region: h.config.Xiaomi.Region,
	}
	missURL, err := xiaomi.ResolveMISSURL(cloudCfg, did, "")
	if err != nil {
		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to resolve MISS URL: %v", err))
		return
	}

	// Connect to camera
	client, err := xiaomi.NewMISSClient(missURL, 30*time.Second)
	if err != nil {
		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to connect to camera: %v", err))
		return
	}
	defer client.Conn.Close()

	// Start speaker
	if err := client.StartSpeaker(); err != nil {
		writeError(w, http.StatusBadGateway, fmt.Sprintf("failed to start speaker: %v", err))
		return
	}
	defer client.StopSpeaker()

	// Upgrade to WebSocket
	upgrader := websocket.Upgrader{
		CheckOrigin: func(r *http.Request) bool { return true },
	}
	conn, err := upgrader.Upgrade(w, r, nil)
	if err != nil {
		logger.Error("WebSocket upgrade failed", "error", err)
		return
	}
	defer conn.Close()

	h.logSuccess(r, "talk.start", "camera", cameraID, "Xiaomi talk session started", nil)
	logger.Info("Xiaomi talk WebSocket connected", "camera_id", cameraID)

	// Send status
	if err := conn.WriteJSON(map[string]any{
		"status":        "connected",
		"camera_id":     cameraID,
		"speaker_codec": client.SpeakerCodec(),
	}); err != nil {
		logger.Error("Failed to send status message", "error", err)
		return
	}

	// Read audio from WebSocket and send to camera
	for {
		messageType, message, err := conn.ReadMessage()
		if err != nil {
			if websocket.IsUnexpectedCloseError(err, websocket.CloseGoingAway, websocket.CloseNormalClosure) {
				logger.Error("Xiaomi talk WebSocket read error", "error", err)
			}
			break
		}

		if messageType == websocket.BinaryMessage {
			if err := client.WriteAudio(client.SpeakerCodec(), message); err != nil {
				logger.Error("Failed to write audio to camera", "camera_id", cameraID, "error", err)
				break
			}
		}
	}

	logger.Info("Xiaomi talk WebSocket disconnected", "camera_id", cameraID)
}