# internal/docsportal/openapi.yaml
openapi: 3.1.0
info:
  title: lalmax-nvr API
  version: dev
  description: |
    lalmax-nvr REST 接口。默认监听 `:9090`。
    REST API for lalmax-nvr. Default listen `:9090`.

    受保护接口使用 **HTTP Basic Auth**。直播播放器无法设置 `Authorization` 时（例如 `EventSource`），可传 `?token=`，值为 base64(`username:password`)。
    Protected routes use **HTTP Basic Auth**. Live players may pass `?token=` as base64(`username:password`) when headers cannot be set.

    错误体固定为 `{"error": "<message>"}`。摄像头启停失败时可能额外返回 `"code"`（如 `CAMERA_NOT_FOUND`）。
    Error bodies are always `{"error": "<message>"}`. Camera start/stop failures may also include `"code"`.

    可浏览文档：NVR `/docs/` · 产品站 `/docs/`。
    Interactive docs: NVR `/docs/` · product site `/docs/`.
  contact:
    url: https://github.com/lalmax-pro/lalmax-nvr
servers:
  - url: http://localhost:9090
    description: 本地 NVR / Local NVR
security:
  - basicAuth: []
...
paths:
...
  /api/cameras/{id}:
    parameters:
      - $ref: "#/components/parameters/CameraID"
    get:
      tags: [Cameras]
      summary: 获取摄像头 · Get camera
      description: |
        返回单路摄像头，含实时录像机状态、错误详情、经纬度等。
        Returns one camera including live recorder status, error detail, and map coordinates.
      operationId: getCamera
      responses:
        "200":
          description: 摄像头（含实时录像状态）/ Camera including live recorder status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Cameras]
      summary: 更新摄像头 · Update camera
      description: |
        部分更新，未出现的字段保持不变。仅当请求里带了非空 `password` 时才改密码。
        Partial update. Omitted fields are left unchanged. Password is only updated when the field is present and non-empty.
      operationId: updateCamera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCamera"
            example:
              name: Front Door (west)
              enabled: true
              longitude: 121.4737
              latitude: 31.2304
      responses:
        "200":
          description: "能重新加载行时返回完整摄像头，否则 `{status: updated}` / Updated camera when the row can be reloaded; otherwise `{status: updated}`"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Camera"
                  - $ref: "#/components/schemas/Status"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Cameras]
      summary: 归档摄像头 · Archive camera
      description: |
        软删除：停止录像机并设 `archived=true`。磁盘上的录像会随摄像头一起归档，不会物理删除。
        Soft-delete. Stops the recorder and sets `archived=true`. Recordings stay on disk and become archived with the camera.
      operationId: deleteCamera
      responses:
        "200":
          description: 已归档 / Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
              example: { status: archived }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: "请求头 `Authorization: Basic base64(username:password)`。默认用户 `admin`。 / `Authorization: Basic base64(username:password)`. Default user is `admin`."
  parameters:
    CameraID:
      in: path
      name: id
      required: true
      schema: { type: string }
      description: "摄像头 id。国标摄像头为 `deviceId:channelId`。 / Camera id (slug). GB28181 cameras use `deviceId:channelId`."
      example: front-door
...
  responses:
    BadRequest:
      description: JSON 无效、缺字段或校验失败 / Invalid JSON, missing field, or failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { error: "invalid request body" }
...
    NotFound:
      description: 资源不存在 / Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { error: "not found" }
  schemas:
    Error:
      title: 错误 · Error
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 可读错误信息 / Human-readable message
        code:
          type: string
          description: 可选机器码（`SETUP_REQUIRED`、`CAMERA_NOT_FOUND` 等） / Optional machine-readable code
    Status:
      title: 状态 · Status
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: 结果标记（`ok` / `started` / `stopped` / `deleted` / `archived` / `acknowledged` 等） / Result token
...
    Camera:
      title: 摄像头 · Camera
      description: |
        摄像头配置与实时录像状态。密码永不回传，只给 `has_password`。
        Camera config plus live recorder status. Password is never returned; only `has_password`.
      type: object
      required: [id, name, protocol, enabled]
      properties:
        id:
          type: string
          description: 摄像头 id / Camera id
        name:
          type: string
          description: 显示名称 / Display name
        protocol:
          type: string
          description: "传输协议：`rtsp` / `http` / `onvif` / `xiaomi` / `gb28181` 及推拉流变体 / Transport: `rtsp`, `http`, `onvif`, `xiaomi`, `gb28181`, plus push/pull variants"
          example: rtsp
        encoding:
          type: string
          enum: [h264, h265, mjpeg, jpeg]
          description: 视频编码 / Video codec
        rtsp_transport:
          type: string
          enum: [tcp, udp]
          description: RTSP 传输层，默认 tcp / RTSP transport, default tcp
        url:
          type: string
          description: 流地址。ONVIF 列表/详情会把 `onvif_endpoint` 填到这里 / Stream URL. For ONVIF this is filled from `onvif_endpoint` in list/get responses
        enabled:
          type: boolean
          description: 是否启用 / Enabled
        description:
          type: string
          description: 备注 / Notes
        location:
          type: string
          description: 安装位置 / Install location
        brand:
          type: string
          description: 品牌 / Brand
        model:
          type: string
          description: 型号 / Model
        serial_number:
          type: string
          description: 序列号 / Serial number
        retention_days:
          type: integer
          description: 本路保留天数。`0` 表示用全局默认，不是永久保存 / Per-camera retention. `0` means use the global default (not keep-forever)
        status:
          type: string
          enum: [recording, paused, stopped, error, reconnecting, offline]
          description: 实时录像机状态 / Live recorder status
        error_type:
          type: string
          nullable: true
          description: 错误分类 / Error class
        error_detail:
          type: string
          nullable: true
          description: 错误详情 / Error detail
        last_seen:
          type: string
          format: date-time
          nullable: true
          description: 最近一次录像时间 / Last recording time
        username:
          type: string
          description: 设备用户名 / Device username
        has_password:
          type: boolean
          description: 是否已保存密码（密码本身不回传）/ Whether a password is stored (never returned)
        onvif_endpoint:
          type: string
          description: ONVIF Device 服务地址 / ONVIF device service URL
        profile_token:
          type: string
          description: ONVIF 主码流 profile / ONVIF main profile token
        profile_name:
          type: string
          description: ONVIF profile 名称 / ONVIF profile name
        stream_encoding:
          type: string
          description: 设备报告的编码（如 H264）/ Device-reported encoding
        audio_enabled:
          type: boolean
          description: 是否采音频 / Audio enabled
        source_type:
          type: string
          description: 接入来源，如 `rtmp_push` / `srt_push` / `whip_push` / `relay_pull` / Ingest origin
        recording_paused:
          type: boolean
          description: 录像是否暂停（拉流仍可能在）/ Recording paused (stream may still be up)
        stream_id:
          type: string
          description: 绑定的 lalmax 流。播放走 `/api/streams/{stream_id}` / Bound lalmax stream. Play via `/api/streams/{stream_id}`
        recording_mode:
          type: string
          enum: [continuous, scheduled, off, event, adaptive]
          description: 只读，来自该流的录像计划；没有计划则为 `off` / Read-only, from the stream plan; `off` if none
          readOnly: true
        activation_state:
          type: string
          enum: [active, pending_activation]
          description: ONVIF 即插即用激活状态 / Plug-and-play activation state
        stable_id:
          type: string
          description: ONVIF 序列号，用于 IP 自愈 / ONVIF serial used for IP self-heal
        sub_stream_url:
          type: string
          description: 子码流 URL / Sub-stream URL
        sub_profile_token:
          type: string
          description: ONVIF 子码流 profile / ONVIF sub-stream profile token
        subnet_hints:
          type: array
          items: { type: string }
          description: IP 重发现用的额外网段 / Extra CIDRs for IP rediscovery
        adaptive:
          type: object
          description: 计划 `mode=adaptive` 时的稀疏录像参数 / Sparse-recording settings when the plan mode is adaptive
          properties:
            timelapse_interval:
              type: string
              description: 平静期写 IDR 的间隔（Go duration，默认 `30s`）/ Calm-state IDR interval (Go duration, default `30s`)
              example: 30s
        archived:
          type: boolean
          description: 是否已归档 / Archived
        archived_at:
          type: string
          format: date-time
          description: 归档时间 / Archived at
        archive_retention_days:
          type: integer
          description: 归档保留天数 / Archive retention days
        longitude:
          type: number
          description: 地图经度（WGS-84）/ WGS-84 longitude for the map view
        latitude:
          type: number
          description: 地图纬度（WGS-84）/ WGS-84 latitude
        created_at:
          type: string
          format: date-time
          description: 创建时间 / Created at
...
    UpdateCamera:
      title: 更新摄像头请求 · Update camera
      description: |
        部分更新，省略的字段不变。
        Partial update; omitted fields stay unchanged.
      type: object
      properties:
        name: { type: string }
        url: { type: string }
        protocol: { type: string }
        encoding: { type: string }
        rtsp_transport: { type: string, enum: [tcp, udp] }
        username: { type: string }
        password: { type: string }
        enabled: { type: boolean }
        description: { type: string }
        location: { type: string }
        brand: { type: string }
        model: { type: string }
        serial_number: { type: string }
        retention_days: { type: integer }
        onvif_endpoint: { type: string }
        profile_token: { type: string }
        stream_encoding: { type: string }
        audio_enabled: { type: boolean }
        sub_stream_url: { type: string }
        sub_profile_token: { type: string }
        subnet_hints:
          type: array
          items: { type: string }
        adaptive:
          type: object
          properties:
            timelapse_interval: { type: string }
        longitude: { type: number }
        latitude: { type: number }
...